# Migrate from npm to pnpm package manager