--- a/.npmrc
+++ b/.npmrc
@@ -1,1 +1,7 @@
-legacy-peer-deps=true
+# pnpm configuration
+auto-install-peers=true
+strict-peer-dependencies=false
+enable-pre-post-scripts=true
+
+# Legacy npm configuration (kept for reference)
+# legacy-peer-deps=true