--- a/package.json
+++ b/package.json
@@ -75,7 +75,7 @@
     "react-resizable-panels": "^2.1.3",
     "recharts": "^2.12.7",
     "sonner": "^1.5.0",
-    "swr": "^2.3.3",
+    "swr": "^2.3.4",
     "tailwind-merge": "^2.5.2",
     "tailwindcss": "3.3.3",
     "tailwindcss-animate": "^1.0.7",